requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
greenlet==1.1.2
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
tzdata==2022.1
